src/main/java/com/gmail/a1ekskosyak/Arrays/TwoDArrayDs.java
package com.gmail.a1ekskosyak.Arrays;

import java.io.IOException;
import java.util.ArrayList;
import java.util.List;
import java.util.Scanner;

public class TwoDArrayDs {

    // Complete the hourglassSum function below.
    static int hourglassSum(int[][] arr) {

        List<Integer> listOfSums = new ArrayList<>();
        for (int i = 0; i < arr.length - 2; i++) {
            for (int j = 0; j < arr[i].length - 2; j++) {
                listOfSums.add(arr[i][j] + arr[i][j + 1] + arr[i][j + 2] + arr[i + 1][j + 1] + arr[i + 2][j] + arr[i + 2][j + 1] + arr[i + 2][j + 2]);
            }
        }
        int maxSum = listOfSums.get(0);
        for (Integer sum : listOfSums) {
            if (sum > maxSum) {
                maxSum = sum;
            }
        }
        return maxSum;
    }

    private static final Scanner scanner = new Scanner(System.in);

    public static void main(String[] args) throws IOException {

        int[][] arr = new int[6][6];

        for (int i = 0; i < 6; i++) {
            String[] arrRowItems = scanner.nextLine().split(" ");
            scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");

            for (int j = 0; j < 6; j++) {
                int arrItem = Integer.parseInt(arrRowItems[j]);
                arr[i][j] = arrItem;
            }
        }

        int result = hourglassSum(arr);

        System.out.println((String.valueOf(result)));

        scanner.close();
    }
}
